chore: move the vega chart to its own file
