+++ src/Chart.tsx
@@ -9,69 +9,24 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { FC, MutableRefObject, forwardRef, useEffect, useMemo, useRef, useState } from 'react';
+import { FC, forwardRef, useEffect, useMemo, useRef, useState } from 'react';
 
 import { EmptyState } from '@components/EmptyState';
 import { LoadingState } from '@components/LoadingState';
-import {
-	DEFAULT_BACKGROUND_COLOR,
-	DEFAULT_COLOR_SCHEME,
-	DEFAULT_LINE_TYPES,
-	DEFAULT_LOCALE,
-	LEGEND_TOOLTIP_DELAY,
-	MARK_ID,
-	SERIES_ID,
-} from '@constants';
+import { DEFAULT_BACKGROUND_COLOR, DEFAULT_COLOR_SCHEME, DEFAULT_LINE_TYPES, DEFAULT_LOCALE } from '@constants';
 import useChartImperativeHandle from '@hooks/useChartImperativeHandle';
 import useChartWidth from '@hooks/useChartWidth';
-import useLegend from '@hooks/useLegend';
-import usePopoverAnchorStyle from '@hooks/usePopoverAnchorStyle';
-import usePopovers, { PopoverDetail } from '@hooks/usePopovers';
 import { useResizeObserver } from '@hooks/useResizeObserver';
-import useSpec from '@hooks/useSpec';
-import useSpecProps from '@hooks/useSpecProps';
-import useTooltips from '@hooks/useTooltips';
 import { getColorValue } from '@specBuilder/specUtils';
-import { getChartConfig } from '@themes/spectrumTheme';
-import {
-	debugLog,
-	getOnMarkClickCallback,
-	getOnMouseInputCallback,
-	sanitizeChartChildren,
-	setSelectedSignals,
-} from '@utils';
-import { VegaChart } from 'VegaChart';
-import { renderToStaticMarkup } from 'react-dom/server';
+import { RscChart } from 'RscChart';
 import { v4 as uuid } from 'uuid';
-import { Item, View } from 'vega';
-import { Handler, Options as TooltipOptions } from 'vega-tooltip';
+import { View } from 'vega';
 
-import {
-	ActionButton,
-	Dialog,
-	DialogTrigger,
-	Provider,
-	View as SpectrumView,
-	defaultTheme,
-} from '@adobe/react-spectrum';
+import { Provider, defaultTheme } from '@adobe/react-spectrum';
 import { Theme } from '@react-types/provider';
 
 import './Chart.css';
-import { ChartData, ChartHandle, ChartProps, Datum, LegendDescription, MarkBounds } from './types';
-
-interface ChartDialogProps {
-	datum: Datum | null;
-	itemName?: string;
-	targetElement: MutableRefObject<HTMLElement | null>;
-	setPopoverState: (isOpen: boolean) => void;
-	popovers: PopoverDetail[];
-}
-
-interface LegendTooltipProps {
-	value: { index: number };
-	descriptions: LegendDescription[];
-	domain: string[];
-}
+import { ChartData, ChartHandle, ChartProps } from './types';
 
 interface PlaceholderContentProps {
 	data: ChartData[];
@@ -116,49 +71,9 @@ export const Chart = forwardRef<ChartHandle, ChartProps>(
 	) => {
 		// uuid is used to make a unique id so there aren't duplicate ids if there is more than one Chart component in the document
 		const chartId = useRef<string>(`rsc-${uuid()}`);
-		const chartView = useRef<View>(); // view returned by vega
-		const selectedData = useRef<Datum | null>(null); // data that is currently selected, get's set on click if a popover exists
-		const selectedDataName = useRef<string>();
-		const selectedDataBounds = useRef<MarkBounds>();
+		// The view returned by vega. This is above RscChart so it can be used for downloading and copying to clipboard.
+		const chartView = useRef<View>();
 		const [containerWidth, setContainerWidth] = useState<number>(0);
-		const popoverAnchorRef = useRef<HTMLDivElement>(null);
-		const [popoverIsOpen, setPopoverIsOpen] = useState<boolean>(false); // tracks the open/close state of the popover
-
-		const sanitizedChildren = sanitizeChartChildren(props.children);
-
-		// THE MAGIC, builds our spec
-		const spec = useSpec({
-			backgroundColor,
-			children: sanitizedChildren,
-			colors,
-			data,
-			description,
-			hiddenSeries,
-			highlightedSeries,
-			symbolShapes,
-			symbolSizes,
-			lineTypes,
-			lineWidths,
-			opacities,
-			colorScheme,
-			title,
-			UNSAFE_vegaSpec,
-		});
-
-		const { controlledHoverSignal, selectedIdSignalName, selectedSeriesSignalName } = useSpecProps(spec);
-		const chartConfig = useMemo(() => getChartConfig(config, colorScheme), [config, colorScheme]);
-
-		useEffect(() => {
-			const tooltipElement = document.getElementById('vg-tooltip-element');
-			if (!tooltipElement) return;
-			// Hide tooltips on all charts when a popover is open
-			tooltipElement.hidden = popoverIsOpen;
-
-			// if the popover is closed, reset the selected data
-			if (!popoverIsOpen) {
-				selectedData.current = null;
-			}
-		}, [popoverIsOpen]);
 
 		useChartImperativeHandle(forwardedRef, { chartView, title });
 
@@ -168,26 +83,6 @@ export const Chart = forwardRef<ChartHandle, ChartProps>(
 		});
 		const chartWidth = useChartWidth(containerWidth, maxWidth, minWidth, width); // calculates the width the vega chart should be
 
-		const {
-			hiddenSeriesState,
-			setHiddenSeries,
-			descriptions: legendDescriptions,
-			isToggleable: legendIsToggleable,
-			onClick: onLegendClick,
-			onMouseOut: onLegendMouseOut,
-			onMouseOver: onLegendMouseOver,
-		} = useLegend(sanitizedChildren); // gets props from the legend if it exists
-
-		const tooltips = useTooltips(sanitizedChildren);
-		const popovers = usePopovers(sanitizedChildren);
-
-		// gets the correct css style to display the anchor in the correct position
-		const targetStyle = usePopoverAnchorStyle(
-			popoverIsOpen,
-			chartView.current,
-			selectedDataBounds.current,
-			padding
-		);
 		const showPlaceholderContent = useMemo(() => Boolean(loading ?? !data.length), [loading, data]);
 		useEffect(() => {
 			// if placeholder content is displayed, clear out the chartview so it can't be downloaded or copied to clipboard
@@ -195,41 +90,6 @@ export const Chart = forwardRef<ChartHandle, ChartProps>(
 				chartView.current = undefined;
 			}
 		}, [showPlaceholderContent]);
-
-		const tooltipConfig: TooltipOptions = { theme: colorScheme };
-
-		if (tooltips.length || legendDescriptions) {
-			tooltipConfig.formatTooltip = (value) => {
-				debugLog(debug, { title: 'Tooltip datum', contents: value });
-				if (value.rscComponentName?.startsWith('legend') && legendDescriptions && 'index' in value) {
-					debugLog(debug, {
-						title: 'Legend descriptions',
-						contents: legendDescriptions,
-					});
-					return renderToStaticMarkup(
-						<LegendTooltip
-							value={value}
-							descriptions={legendDescriptions}
-							// TODO: support multiple legends
-							domain={chartView.current?.scale('legend0Entries').domain()}
-						/>
-					);
-				}
-				// get the correct tooltip to render based on the hovered item
-				const tooltip = tooltips.find((t) => t.name === value.rscComponentName)?.callback;
-				if (tooltip && !('index' in value)) {
-					if (controlledHoverSignal) {
-						chartView.current?.signal(controlledHoverSignal.name, value?.[MARK_ID] ?? null);
-					}
-					return renderToStaticMarkup(
-						<div className="rsc-tooltip" data-testid="rsc-tooltip">
-							{tooltip(value)}
-						</div>
-					);
-				}
-				return '';
-			};
-		}
 
 		if (props.children && UNSAFE_vegaSpec) {
 			throw new Error(
@@ -244,23 +104,6 @@ export const Chart = forwardRef<ChartHandle, ChartProps>(
 			);
 		}
 
-		const signals = useMemo(() => {
-			const signals: Record<string, unknown> = {
-				backgroundColor: getColorValue('gray-50', colorScheme),
-			};
-			if (legendIsToggleable) {
-				signals.hiddenSeries = hiddenSeriesState;
-			}
-			if (selectedIdSignalName) {
-				signals[selectedIdSignalName] = selectedData?.[MARK_ID] ?? null;
-			}
-			if (selectedSeriesSignalName) {
-				signals[selectedSeriesSignalName] = selectedData?.[SERIES_ID] ?? null;
-			}
-
-			return signals;
-		}, [colorScheme, hiddenSeriesState, legendIsToggleable, selectedIdSignalName, selectedSeriesSignalName]);
-
 		return (
 			<Provider
 				colorScheme={colorScheme}
@@ -274,12 +117,6 @@ export const Chart = forwardRef<ChartHandle, ChartProps>(
 					className="rsc-container"
 					style={{ backgroundColor: getColorValue(backgroundColor, colorScheme) }}
 				>
-					<div
-						id={`${chartId.current}-popover-anchor`}
-						data-testid="rsc-popover-anchor"
-						ref={popoverAnchorRef}
-						style={targetStyle}
-					/>
 					{showPlaceholderContent ? (
 						<PlaceholderContent
 							loading={loading}
@@ -288,129 +125,41 @@ export const Chart = forwardRef<ChartHandle, ChartProps>(
 							emptyStateText={emptyStateText}
 						/>
 					) : (
-						<VegaChart
-							spec={spec}
-							config={chartConfig}
+						<RscChart
+							chartView={chartView}
+							chartId={chartId}
 							data={data}
+							backgroundColor={backgroundColor}
+							colors={colors}
+							colorScheme={colorScheme}
+							config={config}
+							description={description}
 							debug={debug}
-							renderer={renderer}
-							width={chartWidth}
 							height={height}
+							hiddenSeries={hiddenSeries}
+							highlightedSeries={highlightedSeries}
+							lineTypes={lineTypes}
+							lineWidths={lineWidths}
 							locale={locale}
+							opacities={opacities}
 							padding={padding}
-							signals={signals}
-							tooltip={tooltipConfig} // legend show/hide relies on this
-							onNewView={(view) => {
-								chartView.current = view;
-								// Add a delay before displaying legend tooltips on hover.
-								let tooltipTimeout: NodeJS.Timeout | undefined;
-								view.tooltip((_, event, item, value) => {
-									const tooltipHandler = new Handler(tooltipConfig);
-									// Cancel delayed tooltips if the mouse moves before the delay is resolved.
-									if (tooltipTimeout) {
-										clearTimeout(tooltipTimeout);
-										tooltipTimeout = undefined;
-									}
-									if (
-										event &&
-										event.type === 'pointermove' &&
-										itemIsLegendItem(item) &&
-										'tooltip' in item
-									) {
-										tooltipTimeout = setTimeout(() => {
-											tooltipHandler.call(this, event, item, value);
-											tooltipTimeout = undefined;
-										}, LEGEND_TOOLTIP_DELAY);
-									} else {
-										tooltipHandler.call(this, event, item, value);
-									}
-								});
-								if (popovers.length || legendIsToggleable || onLegendClick) {
-									if (legendIsToggleable) {
-										view.signal('hiddenSeries', hiddenSeriesState);
-									}
-									setSelectedSignals({
-										selectedData: selectedData.current,
-										selectedIdSignalName,
-										selectedSeriesSignalName,
-										view,
-									});
-									view.addEventListener(
-										'click',
-										getOnMarkClickCallback(
-											chartView,
-											hiddenSeriesState,
-											chartId,
-											selectedData,
-											selectedDataBounds,
-											selectedDataName,
-											setHiddenSeries,
-											legendIsToggleable,
-											onLegendClick
-										)
-									);
-								}
-								view.addEventListener('mouseover', getOnMouseInputCallback(onLegendMouseOver));
-								view.addEventListener('mouseout', getOnMouseInputCallback(onLegendMouseOut));
-								// this will trigger the autosize calculation making sure that everything is correct size
-							}}
-						/>
+							renderer={renderer}
+							symbolShapes={symbolShapes}
+							symbolSizes={symbolSizes}
+							theme={theme}
+							title={title}
+							chartWidth={chartWidth}
+							UNSAFE_vegaSpec={UNSAFE_vegaSpec}
+						>
+							{props.children}
+						</RscChart>
 					)}
-					<ChartDialog
-						datum={selectedData.current}
-						targetElement={popoverAnchorRef}
-						setPopoverState={setPopoverIsOpen}
-						popovers={popovers}
-						itemName={selectedDataName.current}
-					/>
 				</div>
 			</Provider>
 		);
 	}
 );
 Chart.displayName = 'Chart';
-
-const ChartDialog = ({ datum, itemName, targetElement, setPopoverState, popovers }: ChartDialogProps) => {
-	if (!popovers.length) {
-		return <></>;
-	}
-	const popoverDetail = popovers.find((p) => p.name === itemName);
-	const popover = popoverDetail?.callback;
-	const width = popoverDetail?.width;
-	return (
-		<DialogTrigger
-			type="popover"
-			mobileType="tray"
-			targetRef={targetElement}
-			onOpenChange={setPopoverState}
-			placement="top"
-			hideArrow
-		>
-			<ActionButton UNSAFE_style={{ display: 'none' }}>launch chart popover</ActionButton>
-			{(close) => (
-				<Dialog data-testid="rsc-popover" UNSAFE_className="rsc-popover" minWidth="size-1000" width={width}>
-					<SpectrumView gridColumn="1/-1" gridRow="1/-1" margin={12}>
-						{popover && datum && popover(datum, close)}
-					</SpectrumView>
-				</Dialog>
-			)}
-		</DialogTrigger>
-	);
-};
-
-const LegendTooltip: FC<LegendTooltipProps> = ({ value, descriptions, domain }) => {
-	const series = domain[value.index];
-	const description = descriptions.find((d) => d.seriesName === series);
-	if (!description) {
-		return <></>;
-	}
-	return (
-		<div className="rsc-tooltip legend-tooltip" data-testid="rsc-tooltip">
-			<div className="series">{description.title ?? series}</div>
-			<p className="series-description">{description.description}</p>
-		</div>
-	);
-};
 
 const PlaceholderContent: FC<PlaceholderContentProps> = ({ data, emptyStateText, loading, ...layoutProps }) => {
 	if (loading) {
@@ -425,8 +174,4 @@ const PlaceholderContent: FC<PlaceholderContentProps> = ({ data, emptyStateText,
 
 const isValidTheme = (theme: unknown): theme is Theme => {
 	return typeof theme === 'object' && theme !== null && 'light' in theme && 'dark' in theme;
-};
-
-const itemIsLegendItem = (item: Item<unknown>): boolean => {
-	return 'name' in item.mark && typeof item.mark.name === 'string' && item.mark.name.includes('legend');
 };
